+++ aerich/migrate.py
@@ -205,6 +205,8 @@ class Migrate:
                 old_m2m_fields = old_model_describe.get("m2m_fields")
                 new_m2m_fields = new_model_describe.get("m2m_fields")
                 for action, option, change in diff(old_m2m_fields, new_m2m_fields):
+                    if change[0][0] == "db_constraint":
+                        continue
                     table = change[0][1].get("through")
                     if action == "add":
                         add = False
@@ -339,11 +341,14 @@ class Migrate:
                     fk_field = next(
                         filter(lambda x: x.get("name") == new_fk_field_name, new_fk_fields)
                     )
-                    cls._add_operator(
-                        cls._add_fk(model, fk_field, new_models.get(fk_field.get("python_type"))),
-                        upgrade,
-                        fk_m2m=True,
-                    )
+                    if fk_field.get("db_constraint"):
+                        cls._add_operator(
+                            cls._add_fk(
+                                model, fk_field, new_models.get(fk_field.get("python_type"))
+                            ),
+                            upgrade,
+                            fk_m2m=True,
+                        )
                 # drop fk
                 for old_fk_field_name in set(old_fk_fields_name).difference(
                     set(new_fk_fields_name)
@@ -351,13 +356,14 @@ class Migrate:
                     old_fk_field = next(
                         filter(lambda x: x.get("name") == old_fk_field_name, old_fk_fields)
                     )
-                    cls._add_operator(
-                        cls._drop_fk(
-                            model, old_fk_field, old_models.get(old_fk_field.get("python_type"))
-                        ),
-                        upgrade,
-                        fk_m2m=True,
-                    )
+                    if old_fk_field.get("db_constraint"):
+                        cls._add_operator(
+                            cls._drop_fk(
+                                model, old_fk_field, old_models.get(old_fk_field.get("python_type"))
+                            ),
+                            upgrade,
+                            fk_m2m=True,
+                        )
                 # change fields
                 for field_name in set(new_data_fields_name).intersection(set(old_data_fields_name)):
                     old_data_field = next(