+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.5
 
+### 0.5.6
+
+- Add `Command` class. (#148) (#141) (#123) (#106)
+
 ### 0.5.5
 
 - Fix KeyError: 'src_folder' after upgrading aerich to 0.5.4. (#176)